+++ index.html
@@ -226,16 +226,15 @@
               style="text-decoration: underline;" target="_blank">here</a>.
           </b>
         </p>
-        <p style="background: linear-gradient(to bottom right,#f007,#f008);border-radius:var(--roundness);font-size:large;padding:5vmin;
+        <p style="background: linear-gradient(to bottom right,#f006,#f009);border-radius:var(--roundness);font-size:large;padding:5vmin;
           margin:5% 0;
           box-shadow:var(--shadow)">
           To get around the errors on Piped Instances, Please turn on <b>Fetch stream data via Invidious</b> in
           Settings.</p>
         <p
-          style="background: linear-gradient(to bottom right,#0777,#0778);border-radius:var(--roundness);font-size:large;padding:5vmin;box-shadow:var(--shadow)">
-          on IOS there might be a duration bug which can only be fixed via enabling <b>HTTP Live Streaming</b> in
-          Settings.
-        </p>
+          style="background: linear-gradient(to bottom right,#0776,#0779);border-radius:var(--roundness);font-size:large;padding:5vmin;box-shadow:var(--shadow)">
+          On IOS there might be a duration bug which can only be fixed via enabling <b>HTTP Live Streaming</b> in
+          Settings.</p>
       </div>
     </section>
