+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ytify",
   "type": "module",
-  "version": "7.2.3",
+  "version": "7.3",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
@@ -16,13 +16,13 @@
   },
   "devDependencies": {
     "@netlify/edge-functions": "^2.10.0",
-    "@types/node": "^22.5.2",
+    "@types/node": "^22.5.4",
     "@types/sortablejs": "^1.15.8",
     "autoprefixer": "^10.4.20",
     "eruda": "^3.2.3",
     "typescript": "^5.5.4",
-    "vite": "^5.4.2",
-    "vite-plugin-pwa": "^0.20.2",
+    "vite": "^5.4.3",
+    "vite-plugin-pwa": "^0.20.5",
     "vite-plugin-solid": "^2.10.2"
   },
   "browserslist": [
